Refactor ioc to qos and pathioc to pathqos
