+++ pathqos/fs_test.go
@@ -1,4 +1,4 @@
-package pathioc
+package pathqos
 
 import (
 	"sync"
