+++ qos/control_test.go
@@ -1,4 +1,4 @@
-package ioc
+package qos
 
 import (
 	"testing"
