+++ qos/lookup.go
@@ -1,4 +1,4 @@
-package ioc
+package qos
 
 import (
 	"encoding/csv"
