+++ qos/lookup_test.go
@@ -1,4 +1,4 @@
-package ioc
+package qos
 
 import (
 	"strings"
